requirements.txt

# requirements.txt mínimo:
fastapi
uvicorn
psycopg2-binary
pydantic
pytest
pytest-cov
python-dotenv